requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyascii==0.3.2
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.8.2
blis==1.0.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
coremltools==8.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
datasets==3.0.2
dateparser==1.1.8
decorator==5.1.1
diffusers==0.30.3
dill==0.3.8
diskcache==5.6.3
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
g2pkk==0.1.2
grpcio==1.67.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gruut_lang_en==2.0.1
hangul-romanize==0.1.0
huggingface-hub==0.26.1
idna==3.10
importlib_metadata==8.5.0
inflect==7.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llama_cpp_python==0.3.1
llvmlite==0.43.0
macholib==1.16.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
modulegraph==0.19.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.4.2
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
playsound==1.3.0
pooch==1.8.2
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
py2app==0.28.8
pyaml==24.9.0
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyparsing==3.2.0
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
pywhispercpp==1.2.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rumps==0.4.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
trainer==0.0.36
transformers==4.45.2
TTS==0.22.0
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
